fix-bug-28 fix-bug: 4143 【综合模块】编码规则-新增编码规则信息，编码类型无法选择,下拉选出错，物料被校验； 4226 【综合模块】班制-启用的班制数据也可以删除 4068 【班制】新增/编辑界面，开始时间和结束时间前端未显示必填，点击保存时接口提示报错信息 4220 【综合模块】班制-新增班制不填开始和结束时间报错，文案提示异常 4222 【综合模块】班制-新增班制不选择填写班制类型也可以保存 4223 【计划模块】生产日历-新增生产日历信息未启用的班制也会显示 4221 【综合模块】班制-新增班制未跨天开始时间和结束时间相等也可以保存 3928 【4.1.5】备件类型的新增，编码复制问题 3965 【设备模块】备件注册-新增/编辑备件注册，选择备件类型弹窗没有过滤已禁用数据 4200 【接口日志查询】详情界面的返回结果为空

+++ HymsonMES/src/Libraries/Hymson.MES.Services/Services/Plan/PlanShift/PlanShiftService.cs
@@ -143,10 +143,12 @@ namespace Hymson.MES.Services.Services.Plan
                 {
                     //验证数据
                     await _validationDetailRules.ValidateAndThrowAsync(item);
-
+                    int startd = CountDown(item.StartTime);
+                    int endd = CountDown(item.EndTime);
+                    if (startd == endd) { throw new CustomerValidationException(nameof(ErrorCode.MES19510)).WithData("type", item.ShiftType.GetDescription()); }
                     if (!item.IsDaySpan)
                     {
-                        if (CountDown(item.StartTime) > CountDown(item.EndTime))
+                        if (startd > endd)
                         {
                             throw new CustomerValidationException(nameof(ErrorCode.MES19507));
                         }
@@ -161,7 +163,7 @@ namespace Hymson.MES.Services.Services.Plan
                         EndTime = item.EndTime,
                         IsDaySpan = item.IsDaySpan,
                         IsOverTime = item.IsOverTime,
-                        Remark = item.Remark,
+                        Remark = item.Remark ?? string.Empty,
                         Id = IdGenProvider.Instance.CreateId(),
                         CreatedBy = _currentUser.UserName,
                         UpdatedBy = _currentUser.UserName,
@@ -218,6 +220,12 @@ namespace Hymson.MES.Services.Services.Plan
         /// <returns></returns>
         public async Task<int> DeletesAsync(long[] ids)
         {
+            var planShiftEntity = await _planShiftRepository.GetByIdsAsync(ids);
+            if (planShiftEntity.Any(x => x.Status == SysDataStatusEnum.Enable))
+            {
+                throw new CustomerValidationException(nameof(ErrorCode.MES10106));
+            }
+
             //先删除
             await _planShiftRepository.DeletesDetailByIdAsync(ids);
             return await _planShiftRepository.DeletesAsync(new DeleteCommand
@@ -323,11 +331,11 @@ namespace Hymson.MES.Services.Services.Plan
                 throw new CustomerValidationException(nameof(ErrorCode.MES10705));
             }
 
-            if(entity.Status== param.Status)
+            if (entity.Status == param.Status)
             {
                 throw new CustomerValidationException(nameof(ErrorCode.MES10127)).WithData("status", _localizationService.GetResource($"{typeof(SysDataStatusEnum).FullName}.{Enum.GetName(typeof(SysDataStatusEnum), entity.Status)}"));
             }
- 
+
             #endregion
 
             #region 操作数据库
